FIX: notification page structure
